Install Flask API via Ansible
In a previous commit we have installed docker. For the API to be able to run,
we need also python3 and several packages. We install them (if not yet
present).

Afterwards, the `flaskAPIwDocker` directory is coopied to the remote host and
run with `docker-compose`.

Therefore, this ansible script will set up a basic ubuntu system, install
everything necessary for the API to run and start it up with one command only.

Further:
* added typical entries (IDE specific, build stuff, ...) to gitignore (cf.,
* https://www.gitignore.io/api/go,code,linux,macos,windows,eclipse,jetbrains)